requirements.txt
entrypoints==0.3
flake8==3.7.8
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyserial==3.4
